Create new k8s resources (#24)
This commits adds new k8s resources: a `Role` and a `RoleBinding`.
These allow the operator to get only *get* a list of secrets in the
operator's *same* namespace. More details about this choice will be
given in future commits and documentation.
Additionally, other resources have been adapted to the change as
well.

Signed-off-by: Elis Lulja <elulja@cisco.com>